Implementacion Ranking
Manejador ranking agregado, ingresa al usuario cuando se termina la partida y vuelve al menu principal donde puede ingresar a los rankings

+++ juego_Canasta/Assets/Scripts/Player.cs
@@ -10,7 +10,7 @@ public class Player : MonoBehaviour
 
     public Rigidbody2D rb;
 
-    public float puntos;
+    public int puntos;
 
     public bool corriendo;
